css-ui-3_dev/xhtml1print/outline-017.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8" />
<title>CSS Basic User Interface Test: interpolable outline-* properties animation</title>
<link href="http://florian.rivoal.net" rel="author" title="Florian Rivoal" />
<link href="https://drafts.csswg.org/css-ui-3/#propdef-outline-style" rel="help" />
<link href="https://www.w3.org/TR/web-animations-1/#dom-animatable-animate" rel="help" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
@keyframes outline-anim {
  from {
    outline: solid 1px rgb(1,0,0);
    outline-offset: 1px;
  }
  to {
    outline: solid 3px rgb(3,0,0);
    outline-offset: 3px;
  }
}

#test {
  animation: outline-anim 3s -1.5s paused linear;
  outline: solid 1px rgb(1,0,0);
  outline-offset: 1px;
}
</style>
</head><body>
  <div id="test"></div>
  <div id="log"></div>

  <script>
    test(
      function(){
        var test = document.getElementById("test");
        assert_equals(getComputedStyle(test).outlineColor, 'rgb(2, 0, 0)');
      }, "outline-color is animated as a color");
    test(
      function(){
        var test = document.getElementById("test");
        assert_equals(getComputedStyle(test).outlineWidth, '2px');
      }, "outline-width is animated as a length");
    test(
      function(){
        var test = document.getElementById("test");
        assert_equals(getComputedStyle(test).outlineOffset, '2px');
      }, "outline-offset is animated as a length");
</script>

</body></html>
